Fix CI failures on GCC, Clang, macOS, lint, coverage, and sanitizers
  * -Werror builds: GCC and Clang flag conversions MSVC does not.
    detail::shuffle indexes with the iterator's difference type instead
    of casting size_t; replicate_seed drops casts between identical
    64-bit types (-Wuseless-cast); the calibration test's name
    generator no longer shadows gtest's `info` parameter (-Wshadow);
    the benchmarks cast int seeds and int64 ranges explicitly.

  * macOS: Homebrew's libomp is keg-only, so FindOpenMP could not find
    it and configure failed. The CI job exports OpenMP_ROOT from
    `brew --prefix libomp`; the wheel build installs libomp, sets
    OpenMP_ROOT, and raises MACOSX_DEPLOYMENT_TARGET to 14.0 so
    delocate accepts the bundled runtime.

  * clang-format job: Ubuntu 22.04 has no clang-format-19 package, so
    apt failed before any check ran. Install clang-format 19.1.7 from
    PyPI, the version the pre-commit hook pins.

  * clang-tidy job: run-clang-tidy matches file regexes against
    absolute paths, and the unanchored 'mqces/' also matched the
    checkout directory, pulling tests and generated fixtures into the
    run. Anchor the library and CLI patterns at the workspace root.

  * coverage: --coverage was a PRIVATE link option on the static mqces
    library, so executables linking it lacked the gcov runtime and
    failed with undefined __gcov_* symbols. Make it PUBLIC.

  * sanitizers: the ccache step's flow-style mapping contained
    ${{ }}, whose braces closed the mapping early, so the workflow
    never parsed and never ran. Use block style.

+++ benchmarks/bench_classify_end_to_end.cpp
@@ -34,7 +34,8 @@ Setup build_setup(int n_classes, int n_specimens, int d, double epsilon, std::si
   s.known.reserve(static_cast<std::size_t>(n_classes));
   for (int k = 0; k < n_classes; ++k) {
     const double center = static_cast<double>(k) - 0.5 * n_classes;
-    s.known.push_back({"k=" + std::to_string(k), gaussian(n_specimens, d, center, 0x100u + k)});
+    s.known.push_back({"k=" + std::to_string(k),
+                       gaussian(n_specimens, d, center, 0x100u + static_cast<unsigned>(k))});
   }
   s.test = gaussian(n_specimens, d, 0.0, 0xFEED);
   s.opts.weights = mqces::FeatureWeights::Ones(d);
@@ -73,7 +74,7 @@ static void BM_ClassifyClassic_EpsilonSweep(benchmark::State& state) {
   const int n_classes = 19;
   const int n_specimens = 50;
   const int d = 9;
-  const double epsilon = state.range(0) / 100.0;  // 1, 5, 10, 20, 40 %
+  const double epsilon = static_cast<double>(state.range(0)) / 100.0;  // 1, 5, 10, 20, 40 %
   const std::size_t mc = 10;
   auto s = build_setup(n_classes, n_specimens, d, epsilon, mc);
   for (auto _ : state) {
